backend/requirements.txt
python-dotenv~=1.0.1
Flask~=3.0.3
pymongo~=4.10.1
pytest==8.3.3
Flask-Cors~=5.0.0